+++ tests/pointer_transform/pointer_transform_cases/get_bits/expected/get_bits.c
