+++ tests/pointer_transform/pointer_transform_cases/get_bits/input/get_bits.c
