+++ tests/pointer_transform/pointer_transform_cases/nonconst_bound/expected/nonconst_bound.c
