+++ tests/pointer_transform/pointer_transform_cases/nonconst_bound/input/nonconst_bound.c
