+++ tests/pointer_transform/pointer_transform_cases/strchr_wrapper/expected/strchr_wrapper.c
