+++ tests/pointer_transform/pointer_transform_cases/strchr_wrapper/input/strchr_wrapper.c
