+++ tests/pointer_transform/slice_transform_cases/callgraph_propagation/expected/callgraph_propagation.c
