+++ tests/pointer_transform/slice_transform_cases/global_return/expected/global_return.c
