+++ tests/pointer_transform/slice_transform_cases/global_return/input/global_return.c
