+++ tests/pointer_transform/slice_transform_cases/header_fwd_decl/expected/header_fwd_decl.c
