+++ tests/pointer_transform/slice_transform_cases/lo_hi_quicksort/expected/lo_hi_quicksort.c
