+++ tests/pointer_transform/slice_transform_cases/lookaround/expected/lookaround.c
