+++ tests/pointer_transform/slice_transform_cases/lookaround/input/lookaround.c
