+++ tests/pointer_transform/slice_transform_cases/ptr_len_root/expected/ptr_len_root.c
