+++ tests/pointer_transform/slice_transform_cases/ptr_len_root/input/ptr_len_root.c
