+++ tests/pointer_transform/slice_transform_cases/return_index/expected/return_index.c
