+++ tests/pointer_transform/slice_transform_cases/return_index/input/return_index.c
