+++ tests/pointer_transform/slice_transform_cases/singleton_swap/expected/singleton_swap.c
