+++ tests/pointer_transform/slice_transform_cases/singleton_swap/input/singleton_swap.c
